Fix typo of require and move factory bot conf file
